refactor(web): share single-select filter parsing between flows and competition
