+++ apps/web/app/lib/filters.ts
@@ -42,6 +42,37 @@ export function companyListParams(sp: URLSearchParams) {
   };
 }
 
+export interface SingleSelectFilters {
+  sector: string | null;
+  year: string | null;
+  funding: 'all' | 'eu' | 'national';
+  top: number;
+  unknownSector: boolean;
+  unknownYear: boolean;
+}
+
+/**
+ * Single-select explorer filters (sector / year / funding / top) shared by the visual pages
+ * (/flows, /competition). A bogus ?sector or ?year is flagged and dropped from the params, so the
+ * page can show an explicit empty state instead of silently filtering everything out. `years` is the
+ * valid set for the current coverage window.
+ */
+export function singleSelectFilters(sp: URLSearchParams, years: string[]): SingleSelectFilters {
+  const sector = sp.get('sector');
+  const year = sp.get('year');
+  const unknownSector = Boolean(sector) && !KNOWN_SECTORS.has(sector!);
+  const unknownYear = Boolean(year) && !years.includes(year!);
+  const funding = sp.get('funding');
+  return {
+    sector: unknownSector ? null : sector,
+    year: unknownYear ? null : year,
+    funding: funding === 'eu' || funding === 'national' ? funding : 'all',
+    top: sp.get('top') === '50' ? 50 : 20,
+    unknownSector,
+    unknownYear,
+  };
+}
+
 interface SectorFacet {
   value: string;
   label: string;
